requirements.txt
python-uinput
websockets
aiortc
